+++ spec/request/users_spec.rb
@@ -135,77 +135,4 @@ RSpec.describe 'Users controller', type: :request do
     #後で記述
     
   end
-  
-  describe 'アクティベーション' do
-    let!(:user_a) { FactoryBot.create(:user, activated: false) }
-    
-    context 'アクティベーションが成功する' do
-      before do
-        get edit_account_activation_path(user_a.activation_token, email: user_a.email)
-      end
-      
-      it '正しいアクティベートURLにアクセスしてアクティベートされる' do
-        expect(user_a.reload.activated?).to be_truthy
-      end
-      
-      it 'userがログインされ、ユーザー画面が表示される' do
-        expect(response).to redirect_to user_path(user_a)
-      end
-    end
-    
-    context 'アクティベーションが失敗する' do
-      it 'トークンが間違っている' do
-        get edit_account_activation_path(User.new_token, email: user_a.email)
-        expect(user_a.reload.activated?).to be_falsey
-      end
-      
-      it 'emailが間違っている' do
-        get edit_account_activation_path(user_a.activation_token, email: 'wrong@example.com')
-        expect(user_a.reload.activated?).to be_falsey
-      end
-    end
-  end
-    
-  describe 'メール再送機能（resend_email_controller)' do
-    let!(:user_a) { FactoryBot.create(:user, activated: false) }
-    let!(:user_un) { FactoryBot.create(:user, activated: true) }
-    
-    context '未アクティブユーザーのメールアドレスの場合' do
-      before do
-        post resend_emails_path, params: { resend_emails: { email: user_a.email } }
-      end
-      
-      it 'リクエストが成功し、root_urlにリダイレクトされる' do
-        expect(response.status).to eq 302
-        expect(response).to redirect_to root_url
-      end
-    end
-    
-    context 'アクティブ済みユーザーのメールアドレスの場合' do
-      before do
-        post resend_emails_path, params: { resend_emails: { email: user_un.email } }
-      end
-      
-      it 'リクエストが成功し、login_pathにリダイレクトされる' do
-        expect(response.status).to eq 302
-        expect(response).to redirect_to login_path
-      end
-    end
-    
-    context '登録されていないメールアドレスの場合' do
-      before do
-        post resend_emails_path, params: { resend_emails: { email: "wrong@example.com" } }
-      end
-      
-      it 'リクエストが成功し、new_pathにレンダーされる' do
-        expect(response.status).to eq 200
-      end
-      
-      it 'newテンプレートが表示される' do
-        expect(response.body).to include "メールアドレスの確認メールの再送"
-      end
-    end
-  end
-  
-  
 end
